Updating the "Introduction to ART" doc with more information about ART's functionality.
The doc is staged to http://asolovay.mtv:9002/devices/tech/dalvik/art.html

This doc is meant as a partner to the new "Migrating to ART" doc on d.a.c.
( http://ag/469778 ), which has specific information on how developers should
check their apps to avoid ART problems.

(Note: I'd meant to post this directly to source.android.com. As per Clay's
advice, I'll finish the reviews here, check it in to master, then cherrypick it
over to AOSP.)

Change-Id: Ibcb94f81814b2a32e0150a76b99b3abc04ed1e75