Update SELinux documentation.
permissive_or_unconfined() has been removed in master and does not work as
advertised even in 5.x since FORCE_PERMISSIVE_TO_UNCONFINED is always true
in 5.x, so just tell them to always use permissive declarations directly
for domain development.

BOARD_SEPOLICY_UNION has been removed in master, so note that as something
that will not be needed in a future release.

Change-Id: I63bb8f75acd224f0e2fbd18b9cbcc917df06f0de
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>