CDD: Require rotation vector when accel, gyro, magneto available
The rotation vector is now a requirement when a device
includes an accellerometer, a gyroscope sensor and a magentometer
sensor. The requirment has become MANDATORY now as the Android Open
Source project provides the rotation vector sensor implementation
when all 3 base sensors are correctly implemented in the Sensor HAL.

Accelerometer measurements requires at least 12bit resolution.

Bug: 18984395
Bug: 24376867

Change-Id: I6b742954506da72083ec6b1f22c473842e790f94